move: 테스트 쿼리들 파일 폴더 변경
