+++ src/hooks/queries/result/useResultQuery.js
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getTestResults } from "../../api/testResult";
+import { getTestResults } from "../../../api/testResult";
 
 export const useResultQuery = () => {
   return useQuery({
